+++ swagger-adapter/src/main/java/com/vmware/xenon/swagger/SwaggerDescriptorService.java
@@ -13,6 +13,7 @@
 
 package com.vmware.xenon.swagger;
 
+import java.util.EnumSet;
 
 import io.swagger.models.Info;
 
@@ -102,7 +103,7 @@ public class SwaggerDescriptorService extends StatelessService {
 
     @Override
     public void handleGet(Operation get) {
-        Operation op = Operation.createGet(this, "/?includes=ALL&excludes=FACTORY_ITEM");
+        Operation op = Operation.createGet(this, "/");
         op.setCompletion((o, e) -> {
             SwaggerAssembler
                     .create(this)
@@ -115,6 +116,12 @@ public class SwaggerDescriptorService extends StatelessService {
                     .build(get);
         });
 
-        op.sendWith(this);
+        getHost().queryServiceUris(
+                // all services
+                EnumSet.noneOf(ServiceOption.class),
+                true,
+                op,
+                // exclude factory items
+                EnumSet.of(ServiceOption.FACTORY_ITEM));
     }
 }